LINQ ToArray, ToList, ToDictionary tests were updated based on code review
