+++ src/System.Linq/tests/Helpers/TestEnumerable.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace System.Linq.Tests.Helpers
+{
+    public class TestEnumerable<T> : IEnumerable<T>
+    {
+        public T[] Items = new T[0];
+        public TestEnumerable(T[] items) { Items = items; }
+
+        public IEnumerator<T> GetEnumerator() { return ((IEnumerable<T>)Items).GetEnumerator(); }
+        IEnumerator IEnumerable.GetEnumerator() { return Items.GetEnumerator(); }
+    }
+}
