+++ src/System.Linq/tests/System.Linq.Tests.csproj
@@ -21,6 +21,9 @@
     <Compile Include="CachedEnumerator.cs" />
     <Compile Include="EnumerableTests.cs" />
     <Compile Include="EmptyEnumerable.cs" />
+    <Compile Include="Helpers\TestCollection.cs" />
+    <Compile Include="Helpers\TestEnumerable.cs" />
+    <Compile Include="Helpers\TestReadOnlyCollection.cs" />
     <Compile Include="Performance\CastPerf.cs" />
     <Compile Include="Performance\LinqPerformanceCore.cs" />
     <Compile Include="Performance\OrderByThenByPerf.cs" />
@@ -30,7 +33,9 @@
     <Compile Include="Performance\WhereSelectPerf.cs" />
     <Compile Include="Performance\ToArrayToListToDictionaryPerf.cs" />
     <Compile Include="WhereTests.cs" />
-    <Compile Include="ToArrayToListToDictionaryTests.cs" />
+    <Compile Include="ToArrayTests.cs" />
+    <Compile Include="ToDictionaryTests.cs" />
+    <Compile Include="ToListTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\src\System.Linq.csproj">
